requirements.txt
torch
transformers
accelerate
sentencepiece
Pillow
requests
matplotlib
gdown
scikit-image
tensorflow
